Change empty-line convention
If a linkage-module needs a empty line separator, put it at the
beginning rather than at the end of the module.

Add a new "Within-Group"-directive.

All directive-lines will be removed from the linked final program,
except from the main program which must start with a
"Provides"-directive defining a special "main program"-symbol.

To be precise, all the lines up to this line will be removed, and then
the linkage process starts with the directives specified by the
remaining text.