Rename slice fields into "active" and "allocated"
"active" is better than both "length" and "usable". It implies that this
section of the buffer is ready to be accessed without also suggesting that
it already contains valid data.

"length" is too generic and could refer to the "allocated" part of the
buffer as well.

"allocated" is better than "capacity", because the slice can still be
valid (and thus have a "capacity") if it has been statically allocated.
In that case, "allocated" will be 0, indicating that someone else has
allocated the buffer, and we therefore cannot reallocate it.