Unify static and extern linkage variants
The source code now always uses "extern" instead of "static". Without
further actions, this allows the source files to compile as library
functions with external linkage.

However, if

 #define extern static

is defined ahead of the included contents of the source files, the
"extern" declaration will be changed into "static" declarations with
internal linkage.

The only problem are global variables, which must not defined using the
"extern" keyword, but rather without any keyword. In addition, they must
be declared (rather than being defined) in headers file using the
"extern" keyword.

Do the following in such cases:

 /* Source file */
 #ifdef extern
 static
 #endif
 void utility_global;

 /* Header file */
 #ifndef extern
 extern void utility_global;
 #endif

This will define the variable either as static or extern in the source
file, depending on the redefinition of "extern" as a macro.

However, in the header file it will only be declared as a variable with
external linkage which has been defined elsewhere, and this will only be
done if the variable has external linkage. No declarations take place if
the variable uses internal linkage.