Add new dependency "Required-By"
This dependency is "passive", which means it will not pull in any other
module unless one of the modules already pulled in by different
dependencies uses it within a declaration.

"Required-By" could also be called "Before-Group", but then one might
also expect "After-Group" which is however the same as "Requires".

Both "Required-By" and "Requires" refer to groups when used without the
mentioned groups, and refer to individual items named after the group
when used within a within that group group.

Also there is a new idea how to resolve ambiguities: If more than one
module "Provides" the same symbol, this symbol becomes a "weak" symbol.

This means, those modules will not be pulled in automatically. Instead,
they must provide additional unique symbols which cause the module to be
pulled in.