Reconcile documentation with changed '\0'-rules
Also add the restriction that append0() must not be called on a slice
which is already a C string.

Even though it would have been easy to add a runtime check to append0()
in order to check whether the slice is already null-terminated and do
nothing in this case, this would encourage unnecessary calls to
append0() "just to make sure" that an slice is actually a C string.
Those calls would slow down the application for little if any benefit.