Add a document about previous failed ideas
Also change the wording for the case that the data pointer in a slice is
non-NULL but its "allocated" member is 0.

This will now be phrased "non-reallocatable", while both a non-zero
"allocated" as well as a NULL data pointer are considered "allocatable"
or "reallocatable".

The idea is to allow describing a read-only data structure such as a
constant C string literal using a slice.

In that case the "active" member field of the slice structure would
contain the length of the C string including its null terminator, while
the "allocated" member field would be zero.

The same approach would be used for a slice which describes a subsection
of an allocated buffer. Even though the surrounding buffer is
reallocatable, a subsection of it is not. Therefore, the "allocated"
member field would also be zero in this case.