Add collection support
Collections are C source files containing several functions, macros and
global variable definitions which are intended to be used together.

While the fragments/ subdirectory contains indiviual code snippets which
can be #included (or copied) into a new source file, collections require
the developer to copy individually chosen subsections of code out from
the collection file into the new source file.

An advantage of collections versus the snippets in fragments/ is
versioning: It is easy to provide differently specialzed implementations
of the same basic task (group of related definitions which share a
common purpose or collectively provide some sort of service) to the
collections directory. But the fragments/ directory is supposed to only
provide a single coherent version of the implementation of every task
split into multiple snippets; i.e. the latest and "best" implementations
of all tasks.