Add Sections and Groups to documentation
Sections are comments like "I/O functions" which are placed at the
beginning of a group of related functions.

Groups are contiguous ranges of definitions which belong together.

When a definition belongs to a group, 2 things happen:

1. All such definitions are kept inside the same surrounding block.

2. If a definition should "provide" the group, all other group members
   will be placed after this definition.

Without dependencies, definitions will be placed in sorting oder of by
their UUIDs.

This ensures the order of definitions will not change if a definition is
renamed (because its UUID will not be affected by the renaming
operation).